fix: address DGDR controller PR review feedback
- Remove webhook-duplicate validation from validateSpec (image empty check,
  searchStrategy+backend incompatibility) — left to CEL/admission webhook
- Change VRAMMB type from *float64 to *int32
- Remove isOnlineProfiling function (always returned true) and its tests
- Rename checkProfilingJobStatus to isProfilingJobCompleted
- Remove unnecessary r.Get refetch after r.Update in generateAndStoreDGDSpec
- Add clarifying comment for initial '' to Pending phase transition

+++ deploy/operator/api/v1beta1/dynamographdeploymentrequest_types.go
@@ -330,7 +330,7 @@ type HardwareSpec struct {
 
 	// VRAMMB is the VRAM per GPU in MiB.
 	// +optional
-	VRAMMB *float64 `json:"vramMb,omitempty"`
+	VRAMMB *int32 `json:"vramMb,omitempty"`
 
 	// TotalGPUs is the total number of GPUs available in the cluster.
 	// +optional
